requirements.txt
fastapi
uvicorn
pytest
pytest-mock
httpx
python-jose[cryptography]
google-api-python-client
google-auth
